Change client ID type from int to Long
Updated the client ID parameter in ClientController and ClientService from int to Long to support larger identifier values and ensure consistency with repository methods.

+++ src/main/java/com/solarize/solarizeWebBackend/modules/client/ClientService.java
@@ -16,7 +16,7 @@ public class ClientService {
 
     private final ClientRepository REPOSITORY;
 
-    public ClientResponseDTO getClient(int id){
+    public ClientResponseDTO getClient(Long id){
         Optional<Client> client = REPOSITORY.findById(id);
         if(client.isEmpty()) throw new NotFoundException("Client not found.");
         return ClientMapper.of(client.get());
